feat: make the homeparent available globally in the dashboard app to be use on other pages
