+++ dashboard/views/home.py
@@ -1,31 +1,10 @@
-from django.http import Http404
-from django.views.generic import TemplateView
-from django.shortcuts import render, get_object_or_404
+from django.shortcuts import render
 
-from dashboard.models import Course
+from dashboard.util.view import DashboardParentView
 from dashboard.forms.home.update import HomeUpdateForm
-from common.util.views import View
 
 
-class HomeParent(View, TemplateView):
-    def get_context_data(self, *args, **kwargs):
-        context = super().get_context_data(*args, **kwargs)
-
-        course_id = kwargs.pop("course_id", None)
-
-        if not course_id:
-            raise Http404("Course Not found!")
-
-        course = get_object_or_404(Course, pk=course_id)
-
-        context["title"] = f"{course.title}"
-        context["link"] = "course"
-        context["course"] = course
-
-        return context
-
-
-class Home(HomeParent):
+class Home(DashboardParentView):
     template_name = "dashboard/home/detail.html"
 
     def get_context_data(self, *args, **kwargs):
@@ -43,7 +22,7 @@ class Home(HomeParent):
         return render(request, self.template_name, context)
 
 
-class HomeUpdate(HomeParent):
+class HomeUpdate(DashboardParentView):
     template_name = "dashboard/home/update.html"
 
     def get_context_data(self, *args, **kwargs):
