+++ dashboard/views/people.py
@@ -10,15 +10,8 @@ class People(View, TemplateView):
     def get_context_data(self, *args, **kwargs):
         context = super().get_context_data(*args, **kwargs)
 
-        course_id = kwargs.pop("course_id", None)
-
-        course = Course.objects.get(pk=course_id)
-
-        context["title"] = f"{course.title} People"
-        context["link"] = "course"
+        context["title"] = f"{context['title']} People"
         context["sub_link"] = "people"
-
-        context["course"] = course
 
         return context
